backend/pyproject.toml
[project]
name = "ai-lens"
version = "0.1.0"
readme = "README.md"
description = "AgentLens — Agent Training & Production Observability"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "pydantic>2.0",
    "pydantic-settings<3.0.0,>=2.2.1",
    "httpx<1.0.0,>=0.24.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "prometheus-client>=0.17.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "psutil>=5.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
per-file-ignores = { "app/main.py" = ["E402"], "app/mock/builders/*.py" = ["E402"] }

[tool.ruff.format]
quote-style = "double"
